fix outdated bable-eslint issue
There might be a problem with the project dependency tree.
It is likely not a bug in Create React App, but something you need to fix locally.

The react-scripts package provided by Create React App requires a dependency:

  "babel-eslint": "10.0.3"

Don't try to install it manually: your package manager does it automatically.
However, a different version of babel-eslint was detected higher up in the tree:

  /example-consumer-cypress/node_modules/babel-eslint (version: 10.1.0)

Manually installing incompatible versions is known to cause hard-to-debug issues.